build: Add PHP v5.3 and patch v5.4, remove Windows SDK, add test config
Windows SDK support has been removed due to the nature of the PHP
build process; Windows SDK is not possible. In order to create a
simple configuration process for building the driver and creating a
process to run tests in Windows, zlib is being enabled by default for
the cpp-driver and in turn the php-driver (and OpenSSL). This allows
for enabled the zip extension during the PHP build process to work
better with composer for installing dependencies.

Driver builds and additional testing configuration builds have been
split up to allow for quicker builds when just the driver is wanted.
Test configuration builds are disabled by default but can be enabled
with --ENABLE-TEST-CONFIGURATION.