win: Correct build issues with VS2010 for PHP v5.3
This commit allows for the dirver to build against PHP v5.3. A small
fix in the build script was also made to ensure PHP dependencies were
copied correctly and that cpp-driver options were targeted as they
should be.

The cpp-driver submodule was also updated to fix an issue when
building with Boost atomic (PHP driver build script forces this
setting for VS2010 builds).